backend/services/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "oziebot-api"
version = "0.1.0"
description = "Oziebot HTTP API (tenants, auth, billing webhooks)"
requires-python = ">=3.12"
dependencies = [
  "oziebot-domain",
  "oziebot-common",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic-settings>=2.0",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "alembic>=1.13",
  "redis>=5.0",
  "passlib>=1.7.4",
  "python-jose[cryptography]>=3.3.0",
  "email-validator>=2.0",
  "stripe>=11.0",
  "cryptography>=42.0",
  "httpx>=0.27",
  "PyJWT>=2.8",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["src/oziebot_api"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "../strategy-engine/src"]
